Move BigQuery call for /user_reports off mainthread
While the actual query was already async, this ensures we aren't
trying to transfer JS arrays/objects between threads unecessarily.
